Improved logging, fixed typos, organized file structure main.py will run without errors
