+++ .gitignore
@@ -1,0 +1,3 @@
+*.pyc
+*/debug.log
+*/__pycache__/*
